LayoutTests/crypto/webkitSubtle/aes-cbc-wrong-key-class-expected.txt

CONSOLE MESSAGE: line 22: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test calling AES-CBC encrypt with a HMAC key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw HMAC key from string literal...
Using the key to encrypt plaintext...
PASS crypto.webkitSubtle.encrypt({name: "aes-cbc", iv: iv}, key, plaintext) threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
